feat(events): move EventData and related classes
